feat: tests working
