+++ src/app/core/components/footer/footer.component.spec.ts
@@ -8,7 +8,9 @@ describe('FooterComponent', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      declarations: [FooterComponent]
+      imports: [
+        FooterComponent,
+      ],
     });
     fixture = TestBed.createComponent(FooterComponent);
     component = fixture.componentInstance;
